Removed the last of the Visual Studio files.
The rationale for having it was that with videoInput, building OpenTLD
under Visual Studio 2010 required that we put
"atlthunk.lib;libcmt.lib" in the "Ignore specific default libraries"
field. However, since CMake keeps rebuilding this file, we need to
manually change that field anyway, so having the file around for that
has proven too much hassle.